NormalyzerResults {NormalyzerDE}	R Documentation

Representation of the results from performing normalization over a dataset

Description

It is linked to a NormalyzerDataset instance representing the raw data which has been processed. After performing evaluation it also links to an instance of NormalyzerEvaluationResults representing the results from the evaluation.

Usage

NormalyzerResults(nds)

NormalyzerResults(nds)

Arguments

nds
NormalyzerDataset object

Value

nr Prepared NormalyzerResults object

Slots

normalizations

SummarizedExperiment object containing calculated normalization results

nds

Normalyzer dataset representing run data

ner

Normalyzer evaluation results for running extended normalizations

Examples

data(example_summarized_experiment)
normObj <- getVerifiedNormalyzerObject("job_name", example_summarized_experiment)
emptyNormResults <- NormalyzerResults(normObj)

[Package NormalyzerDE version 1.12.0 Index]