calculatePercentageAvgDiffInMat {NormalyzerDE}	R Documentation

General function for calculating percentage difference of average column means in matrix

Description

General function for calculating percentage difference of average column means in matrix

Usage

calculatePercentageAvgDiffInMat(targetMat)

Arguments

targetMat
Matrix for which column means should be compared

Value

percDiffVector Vector with percentage difference, where first element always will be 100

[Package NormalyzerDE version 1.12.0 Index]