example_data {NormalyzerDE}	R Documentation

Small example dataset used to demonstrate code consistency in testing and as dummy data in the vignette.

Description

Small example dataset used to demonstrate code consistency in testing and as dummy data in the vignette.

Usage

example_data

Format

A data frame containing annotation and expression data

[Package NormalyzerDE version 1.12.0 Index]