filterLowRep {NormalyzerDE}	R Documentation

Filter rows with lower than given number of replicates for any condition

Description

Filter rows with lower than given number of replicates for any condition

Usage

filterLowRep(df, groups, leastRep = 2)

Arguments

df
Dataframe with expression data to filter

groups
Condition groups header

leastRep
Minimum number of replicates in each group to retain

Value

collDesignDf Reduced design matrix

[Package NormalyzerDE version 1.12.0 Index]