getReplicateSortedData {NormalyzerDE}	R Documentation

Get dataframe with raw data column sorted on replicates

Description

Get dataframe with raw data column sorted on replicates

Usage

getReplicateSortedData(rawDataOnly, groups)

Arguments

rawDataOnly
Dataframe with unparsed input data matrix.

groups
Vector containing condition levels.

Value

rawData sorted on replicate

[Package NormalyzerDE version 1.12.0 Index]