plotContrastPCA {NormalyzerDE}	R Documentation

Show in a PCA plot what samples are compared in statistical contrast This is useful to understand what conditions are compared and for checking for outliers in the contrast

Description

Show in a PCA plot what samples are compared in statistical contrast This is useful to understand what conditions are compared and for checking for outliers in the contrast

Usage

plotContrastPCA(nst, jobName, currentLayout, pageno, pcs = c(1, 2))

Arguments

nst
NormalyzerDE statistics object.

jobName
Name of processing run.

currentLayout
Layout used for document.

pageno
Current page number.

pcs
Principal components to show.

Value

None

[Package NormalyzerDE version 1.12.0 Index]