plotQQ {NormalyzerDE}	R Documentation

Produces page showing QQ-plots for the first sample for each normalization method. This plot can be used to assess whether the data follows a normal distribution.

Description

Produces page showing QQ-plots for the first sample for each normalization method. This plot can be used to assess whether the data follows a normal distribution.

Usage

plotQQ(nr, currentLayout, pageno)

Arguments

nr
Normalyzer results object.

currentLayout
Layout used for document.

pageno
Current page number.

Value

None

[Package NormalyzerDE version 1.12.0 Index]