plotReplicateVarAndStableVariables {NormalyzerDE}	R Documentation

Write figures displaying pooled coefficient of variance, median absolute deviation and pooled estimate of variance percentage compared to log2-transformed and stable variables plot displaying CV of stable variables against pooled CV measure. The stable variables are calculated by an ANOVA comparison across sample conditions and selecting features with the least clear difference.

Description

Write figures displaying pooled coefficient of variance, median absolute deviation and pooled estimate of variance percentage compared to log2-transformed and stable variables plot displaying CV of stable variables against pooled CV measure. The stable variables are calculated by an ANOVA comparison across sample conditions and selecting features with the least clear difference.

Usage

plotReplicateVarAndStableVariables(nr, currentLayout, pageno)

Arguments

nr
Normalyzer results object.

currentLayout
Layout used for document.

pageno
Current page number.

Value

None

[Package NormalyzerDE version 1.12.0 Index]