printPlots {NormalyzerDE}	R Documentation

Generate PDF grid page filling it with provided list of plots

Description

Generate PDF grid page filling it with provided list of plots

Usage

printPlots(plotlist, plotname, pageno, jobname, currentLayout)

Arguments

plotlist
List of target plots to display.

plotname
List of names corresponding to the provided plot list.

pageno
Current page number.

jobname
Name of ongoing job.

currentLayout
Custom viewport layout.

Value

None

[Package NormalyzerDE version 1.12.0 Index]