setupJobDir {NormalyzerDE}	R Documentation

Create empty directory for run

Description

Creates a directory at provided path named to the jobname.

Usage

setupJobDir(jobName, outputDir)

Arguments

jobName
Name of the run.

outputDir
Path to directory where to create the output directory.

Value

Path to newly created directory.

Examples

setupJobDir("job_name", "path/to/outdir")

[Package NormalyzerDE version 1.12.0 Index]