setupPlotting {NormalyzerDE}	R Documentation

Setup PDF report settings by initializing the color palette, format for the PDF report and the graphical device

Description

Setup PDF report settings by initializing the color palette, format for the PDF report and the graphical device

Usage

setupPlotting(currentJob, jobDir, suffix)

Arguments

currentJob
Name of current run.

jobDir
Path to output directory for run.

suffix
Text to add to output filename.

Value

None

[Package NormalyzerDE version 1.12.0 Index]