verifyDesignMatrix {NormalyzerDE}	R Documentation

Verify that design matrix setup matches the data matrix

Description

Verify that design matrix setup matches the data matrix

Usage

verifyDesignMatrix(fullMatrix, designMatrix, sampleCol)

Arguments

fullMatrix
Dataframe with input data.

designMatrix
Dataframe with design setup.

sampleCol
Column in design matrix containing sample IDs.

Value

None

[Package NormalyzerDE version 1.12.0 Index]