verifyValidNumbers {NormalyzerDE}	R Documentation

Verify that input fields conform to the expected formats

Description

Verify that input fields conform to the expected formats

Usage

verifyValidNumbers(rawDataOnly, groups, noLogTransform = FALSE, quiet = FALSE)

Arguments

rawDataOnly
Dataframe with input data.

groups
Condition levels for comparisons.

Value

None

[Package NormalyzerDE version 1.12.0 Index]