NormalyzerEvaluationResults {NormalyzerDE}	R Documentation

Representation of evaluation results by calculating performance measures for an an NormalyzerResults instance

Description

Contains the resulting information from the processing which subsequently can be used to generate the quality assessment report.

Usage

NormalyzerEvaluationResults(nr)

NormalyzerEvaluationResults(nr)

Arguments

nr
NormalyzerResults object

Value

nds Generated NormalyzerEvaluationResults instance

Slots

avgcvmem

Average coefficient of variance per method

avgcvmempdiff

Percentage difference of mean coefficient of variance compared to log2-transformed data

featureCVPerMethod

CV calculated per feature and normalization method.

avgmadmem

Average median absolute deviation

avgmadmempdiff

Percentage difference of median absolute deviation compared to log2-transformed data

avgvarmem

Average variance per method

avgvarmempdiff

Percentage difference of mean variance compared to log2-transformed data

lowVarFeaturesCVs

List of 5 for log2-transformed data

lowVarFeaturesCVsPercDiff

Coefficient of variance for least variable entries

anovaP

ANOVA calculated p-values

repCorPear

Within group Pearson correlations

repCorSpear

Within group Spearman correlations

Examples

data(example_summarized_experiment)
normObj <- getVerifiedNormalyzerObject("job_name", example_summarized_experiment)
normResults <- normMethods(normObj)
normEval <- NormalyzerEvaluationResults(normResults)

[Package NormalyzerDE version 1.12.0 Index]