writeNormalizedDatasets {NormalyzerDE}	R Documentation

Write normalization matrices to file

Description

Outputs each of the normalized datasets to the specified directory.

Usage

writeNormalizedDatasets(
  nr,
  jobdir,
  includePairwiseComparisons = FALSE,
  includeCvCol = FALSE,
  includeAnovaP = FALSE,
  normSuffix = "-normalized.txt",
  rawdataName = "submitted_rawdata.txt"
)

Arguments

nr
Results object.

jobdir
Path to output directory.

includePairwiseComparisons
Include p-values for pairwise comparisons.

includeCvCol
Include CV column in output.

includeAnovaP
Include ANOVA p-value in output.

normSuffix
String used to name output together with normalization names.

rawdataName
Name of output raw data file.

Value

None

Examples

data(example_summarized_experiment)
normObj <- getVerifiedNormalyzerObject("job_name", example_summarized_experiment)
normResults <- normMethods(normObj)
normResultsWithEval <- analyzeNormalizations(normResults)
outputDir <- tempdir()
writeNormalizedDatasets(normResultsWithEval, outputDir)

[Package NormalyzerDE version 1.12.0 Index]